% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{R/make_codons.R
Create codons from DNA
input: dna, s=1
return: codons}
\usage{
mk_codons(dna, s = 1)
}
\description{
R/make_codons.R
Create codons from DNA
input: dna, s=1
return: codons
}
