% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna_function.R
\name{random_dna}
\alias{random_dna}
\title{R/random_dna.R
Sample random string of nucleotides of size l
input: l, length
return: nucleotides}
\usage{
random_dna(l)
}
\description{
R/random_dna.R
Sample random string of nucleotides of size l
input: l, length
return: nucleotides
}
