% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_codons_to_aa.R
\name{RNA_codons_to_aa}
\alias{RNA_codons_to_aa}
\title{This function converts RNA-codons to amino acids}
\usage{
RNA_codons_to_aa(codons)
}
\description{
This function converts RNA-codons to amino acids
}
